Rename to file_upload and care_type
